initial commit for dual_form templatetag
introduced new templatetag called 'dual_form'. It's supposed to be used
like this:

{% dual_form field_a,field_b[,...] from form as left|right|fullrow %}

This templatetag will be replaced with a list of divs for each form.
Each div will be decorated with a class name depending on the last
argument, which will make the field align to the left or right column or
to be a fullrow field.

Additional features are that helptexts are displayed using tipsy.js,
required fields' labels have an asterisk (*) appended, error messages
are displayed properly and fields with validation errors are decorated
with the 'tx-form-error' class name.