Return count of deleted strings from save2db, too.
The save2db function already returned the number of strings added and
that of those deleted. We extend the return result to include the number
of strings deleted, too.

This is needed to push some of the actions out of the save2db, which is
necessary for the correct handling of the transaction, among other
things.