signal to check if changing project's wordcount is allowed
This signal will be raised whenever an action that attempts to change
the project's wordcount happens.

The signal contains the project and a reference to a list called
'errors'. If the signal handler decides that the wordcount change should
not be allowed, it appends the reason to the 'errors' list.